fix: fix bitset and make bitset size fix

+++ pkg/core/record/manager.go
@@ -3,6 +3,7 @@ package record
 import (
 	"github.com/pigeonligh/stupid-base/pkg/core/storage"
 	"github.com/pigeonligh/stupid-base/pkg/core/types"
+	log "github.com/pigeonligh/stupid-base/pkg/logutil"
 )
 
 type Manager struct {
@@ -21,7 +22,7 @@ func init() {
 	}
 }
 
-func (m *Manager) CreateFile(filename string, recordSize uint32) error {
+func (m *Manager) CreateFile(filename string, recordSize int) error {
 	var err error
 	if err = m.storage.CreateFile(filename); err != nil {
 		return err
@@ -42,10 +43,9 @@ func (m *Manager) CreateFile(filename string, recordSize uint32) error {
 	header.RecordNum = 0
 	header.RecordPerPage = recordPerPage(recordSize)
 	header.SlotMapSize = bitMapSize(header.RecordPerPage)
-	header.SizeOfHeader = header.SlotMapSize + 4 // equals to sizeof(PageNum)
 
-	header.Pages = 1
-	header.FirstFree = 0
+	header.FileHeaderPage.Pages = 1
+	header.FileHeaderPage.FirstFree = 0
 
 	if err = fileHandle.MarkDirty(pageHandle.Page); err != nil {
 		return err
@@ -57,34 +57,31 @@ func (m *Manager) CreateFile(filename string, recordSize uint32) error {
 }
 
 func (m *Manager) DestroyFile(filename string) error {
+	if _, found := m.files[filename]; found {
+		// TODO: should return warn for open file which is opened
+		log.V(log.RecordLevel).Warningf("DestroyFile failed: %v, file opened!", filename)
+		return nil
+	}
 	if err := m.storage.DestroyFile(filename); err != nil {
 		return err
 	}
+	log.V(log.RecordLevel).Infof("DestroyFile succeeded: %v", filename)
 	return nil
 }
 
 func (m *Manager) OpenFile(filename string) (*FileHandle, error) {
 
 	if file, found := m.files[filename]; found {
+		// TODO: should return warn for open file which is opened
+		log.V(log.RecordLevel).Infof("OpenFile: %v has already opened! ", filename)
 		return file, nil
 	}
-
-	storageFH, err := m.storage.OpenFile(filename)
-	if err != nil {
-		return nil, err
-	}
-	pageHandle, err := storageFH.GetPage(0)
-	if err != nil {
-		return nil, err
-	}
-
 	// RM_FileHandle
-	file := newFileHandle(filename, pageHandle)
-
-	if err = storageFH.UnpinPage(pageHandle.Page); err != nil {
+	file, err := newFileHandle(filename)
+	if err != nil {
 		return nil, err
 	}
-
+	log.V(log.RecordLevel).Infof("OpenFile succeeded: %v", filename)
 	m.files[filename] = file
 	return file, nil
 }
